feat(scuola): refactor anno scolastico routes and controllers for improved structure
